+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.4 (unreleased)
 
 - Add `record_ipv4`, `record_ipv6`, `record_mpls` and `sflow_options` attributes to `iosxr_flow_monitor_map` resource and data source
+- Add `set_overload_bit` attributes to `iosxr_router_isis` resource and data source
 
 ## 0.2.3